Use small action names for domain actions
The "Make Payment" implies a use case, not a domain action.
When we make a payment, we assume that we will take some data,
payment amount, current balance, etc., and create a new payment
associated with the account, not simply add an existing payment
to an account. This Add function will be part of the implementation
of a MakePayment use case function.